fix(test): the fault-pair checkpoint scan compiles under GCC
std::max(best, std::strtoull(...)) fails deduction on Linux: strtoull
returns unsigned long long and GCC's uint64_t is unsigned long, two
different types. Apple clang defines uint64_t AS unsigned long long, so
the host build compiled it and the break only surfaced in CI - the
known host-vs-Linux class, same member as the previous std::max case.
Reproduced and the cast verified against g++ in the clink-build image
before pushing: the old form fails with CI's exact error, the cast
compiles.

+++ tests/integration/test_fault_pairs.cpp
@@ -150,7 +150,12 @@ std::uint64_t latest_completed(const std::filesystem::path& ckpt_root) {
         }
         const auto name = it->path().filename().string();
         if (name.rfind("COMPLETED-", 0) == 0) {
-            best = std::max(best, std::strtoull(name.c_str() + 10, nullptr, 10));
+            // Explicit cast: strtoull returns unsigned long long, and under
+            // GCC uint64_t is unsigned long - different types, so the bare
+            // std::max fails deduction there (compiles fine on Apple clang,
+            // where the two coincide).
+            best = std::max(
+                best, static_cast<std::uint64_t>(std::strtoull(name.c_str() + 10, nullptr, 10)));
         }
     }
     return best;
